+++ libportal/wallpaper.c
@@ -42,7 +42,7 @@ typedef struct {
   guint signal_id;
   GTask *task;
   char *request_path;
-  guint cancelled_id;
+  gulong cancelled_id;
 } WallpaperCall;
 
 static void
